pyproject.toml
[tool.poetry]
name = "em-fastapi"
version = "0.1.0"
description = ""
authors = ["Nikita Smykov <132088678+Apicqq@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
pydantic-settings = "^2.6.1"
sqlalchemy = "^2.0.36"
uvicorn = "^0.32.1"
asyncpg = "^0.30.0"
fastapi-pagination = "^0.12.32"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
